# docs/swagger.yaml
swagger: "2.0"
info:
  description: "API for managing companies"
  version: "1.0"
  title: "Company API"
basePath: "/v1"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
  /auth/register:
    post:
      summary: "Register a new company and manager"
      tags:
        - auth
      operationId: "registerCompany"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/RegisterCompanyRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            properties:
              status:
                description: 'Status of operation'
                type: string
                example: success
...
  /companies/{id}:
    get:
      summary: "Get a company by ID"
      tags:
          - companies
      operationId: "getCompany"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cd8d8-4edf-7584-8042-4d9adefa2641"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Company"
    patch:
      summary: "Update a company by ID"
      tags:
          - companies
      operationId: "patchCompany"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cd8d8-4edf-7584-8042-4d9adefa2641"
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/PatchCompanyRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Company"
      requestBody:
        content:
          application/json:
            example:
              name: "Updated Company"
              short: "UpCo"
              pib: "987654321"
              addressId: "018cfd18-f492-7ade-803a-b414a0a8afac"
              mn: "123456789"
              phone: "+987654321"
              email: "updatedcompany@example.com"

    delete:
      summary: "Delete a company by ID"
      tags:
          - companies
      operationId: "deleteCompany"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Company"
...
  /repairs:
    get:
      summary: "Search all repairs"
      tags:
        - repairs
      operationId: "searchAllRepairs"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Repair"

    post:
      summary: "Create a new repair"
      tags:
        - repairs
      operationId: "createRepair"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateRepairRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Repair"          
  /repairs/{id}:
    get:
      summary: "Get a repair by ID"
      tags:
        - repairs
      operationId: "getRepair"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Repair"
    patch:
      summary: "Update a repair by ID"
      tags:
        - repairs

      operationId: "patchRepair"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/PatchRepairRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Repair"
      requestBody:
        content:
          application/json:
            example:
              addressId: "018cfd18-f492-7ade-803a-b414a0a8afac"
              warrantId: "018cfd18-f674-7d5e-8051-a4dea1acbe79"
              expectedTime: "2023-12-31T12:00:00Z"
              order: 1
              
    delete:
      summary: "Delete a repair by ID"
      tags:
        - repairs
      operationId: "deleteRepair"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Repair"
...
  /transportCosts:
    get:
      summary: "Search all transportCosts"
      tags:
        - transport costs
      operationId: "searchAllTransportCosts"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TransportCosts"

    post:
      summary: "Create a new transportCosts"
      tags:
        - transport costs
      operationId: "createTransportCosts"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateTransportCostsRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/TransportCosts"          
  /transportCosts/{id}:
    get:
      summary: "Get a repair by ID"
      tags:
        - transport costs

      operationId: "getTransportCosts"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cfd18-f6f0-7113-8059-4afc7fa2691d"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/TransportCosts"
    patch:
      summary: "Update a transportCosts by ID"
      tags:
        - transport costs
      operationId: "patchTransportCosts"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cfd18-f6f0-7113-8059-4afc7fa2691d"
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/PatchTransportCostsRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/TransportCosts"
      requestBody:
        content:
          application/json:
            example:
              addressId: "018cfd18-f492-7ade-803a-b414a0a8afac"
              warrantId: "018cfd18-f674-7d5e-8051-a4dea1acbe79"
              expectedTime: "2023-12-31 12:00:00"
              order: 1
              
    delete:
      summary: "Delete a transportCosts by ID"
      tags:
        - transport costs
      operationId: "deleteTransportCosts"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cfd18-f6f0-7113-8059-4afc7fa2691d"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/TransportCosts"          
  /driverCategories:
    get:
      summary: "Search all driver categories"
      tags:
        - driver categories
      operationId: "searchAllDriverCategories"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DriverCategory"

    post:
      summary: "Create a new driver category"
      tags:
        - driver categories

      operationId: "createDriverCategory"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateDriverCategoryRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/DriverCategory"          
  /driverCategories/{id}:
    get:
      summary: "Get a driver category by ID"
      tags:
        - driver categories

      operationId: "getDriverCategory"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cf89f-28ce-7215-8008-b7886d2dfa05"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/DriverCategory"
    patch:
      summary: "Update a driver category by ID"
      tags:
        - driver categories

      operationId: "patchDriverCategory"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cf89f-28ce-7215-8008-b7886d2dfa05"
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/PatchDriverCategoryRequest"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/DriverCategory"
      requestBody:
        content:
          application/json:
            example:
              category: "Updated Category"
              issued: "2022-01-01"
              expired: "2023-01-01"

    delete:
      summary: "Delete a driver category by ID"
      tags:
        - driver categories

      operationId: "deleteDriverCategory"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "string"
          default: "018cf8f1-7279-7cb8-800d-75c9227a7544"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/DriverCategory"
...
  /routes/:
    get:
      summary: Get all routes
      tags:
        - routes
      operationId: searchAllRoutes
      produces:
        - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            type: array
            items:
              $ref: "#/definitions/Route"

    post:
      summary: Create a route
      tags:
        - routes
      operationId: createRoute
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Route object
          required: true
          schema:
            $ref: "#/definitions/CreateRouteRequest"
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Route"
  /routes/{id}:
    get:
      summary: Get a route by ID
      tags:
        - routes
      operationId: getRoute
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the warrant to get
          required: true
          type: string
          default: "018cfd25-a957-76ec-8094-4807812a277b"
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Route"

    patch:
      summary: Update a route by ID
      tags:
        - routes
      operationId: patchRoute
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the route to update
          required: true
          type: string
          default: "018cfd25-a957-76ec-8094-4807812a277b"
        - in: body
          name: body
          description: PatchRoute object
          required: true
          schema:
            $ref: "#/definitions/PatchRouteRequest"
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Route"

    delete:
      summary: Delete a route by ID
      tags:
        - routes
      operationId: deleteRoute
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the route to delete
          required: true
          type: string
          default: "018cfd25-a957-76ec-8094-4807812a277b"
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Route"
...
definitions:
  Company:
    type: "object"
    properties:
      id:
        type: "string"
        example: "12345"
      name:
        type: "string"
        example: "Example Company"
      short:
        type: "string"
        example: "ExCo"
      pib:
        type: "string"
        example: "123456789"
      addressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      address:
        $ref: "#/definitions/Address"
      mn:
        type: "string"
        example: "987654321"
      phone:
        type: "string"
        example: "+123456789"
      email:
        type: "string"
        example: "info@example.com"
      employees:
        type: array
        items:
          $ref: "#/definitions/UserCompany"
      meta:
        type: object
        additionalProperties: true
        example: { "key": "value" }
  PatchCompanyRequest:
    type: "object"
    properties:
      short:
        type: "string"
        example: "UpCo"
      addressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      phone:
        type: "string"
        example: "+987654321"
      email:
        type: "string"
        example: "updatedcompany@example.com"

  # CreateCompanyRequest:
  #   type: "object"
  #   properties:
  #     name:
  #       type: "string"
  #       example: "New Company"
  #     short:
  #       type: "string"
  #       example: "NewCo"
  #     pib:
  #       type: "string"
  #       example: "987654321"
  #     addressId:
  #       type: "string"
  #       example: "018cfd18-f492-7ade-803a-b414a0a8afac"
  #     mn:
  #       type: "string"
  #       example: "123456789"
  #     phone:
  #       type: "string"
  #       example: "+987654321"
  #     email:
  #       type: "string"
  #       example: "newcompany@example.com"
  RegisterCompanyRequest:
    type: "object"
    properties:
      companyAddressCity:
        type: "string"
        example: "Beograd"
      companyAddressStreet:
        type: "string"
        example: "Gagarina"
      companyAddressStreetNo:
        type: "string"
        example: "204a"
      companyAddressName:
        type: "string"
        example: "Beogradjanka"
      companyName:
        type: "string"
        example: "Naftna Industrija Srbije"
      companyShort:
        type: "string"
        example: "NIS"
      pib:
        type: "string"
        example: "12231222123"  
      companyAddressID:
        type: "string"
        example: "wdjkwskjsa-2dsadkdsad-dsadsdas"
      companyMn:
        type: "string"
        example: "231323131"
      companyPhone:
        type: "string"
        example: "+3812312131"
      companyEmail:
        type: "string"
        example: "offilce@oblak.com"
      companyMeta:
        type: "object"
        example:
          status: submitted
          addressId: asdsadsa-sadasdsad-adsdsad
      userAddressCity:
        type: "string"
        example: "Zemun"
      userAddressStreet:
        type: "string"
        example: "Dusanova"
      userAddressStreetNo:
        type: "string"
        example: "123a"
      userAddressName:
        type: "string"
        example: "zemunska posta"
      userAddressID:
        type: "string"
        example: "wdjkwskjsa-2dsadkdsad-dsadsdas"
      userEmail:
        type: "string"
        example: "user@random.com"
      userPassword:
        type: "string"
        example: "asdfasdf"
      userFirstName:
        type: "string"
        example: "Milos"
      userLastName:
        type: "string"
        example: "Bokovic"
      userMn:
        type: string
        example: "1231323"
      userBirthplace:
        type: string
        example: "Beograd"
      isDriver:
        type: bool
        example: False
      licenceNumber:
        type: string
        example: 1223213123a
      licenceSerialNumber: 
        type: string
        example: 332d2dddsds
      licenceExpiry:
        type: date
        example: 2024-01-02
      licenceIssued:
        type: date
        example: 2023-01-02
      licenceAuthority:
        type: string
        example: PU Novi Beograds
...
  Address:
    type: "object"
    properties:
      id:
        type: "string"
        required: true
        example: "98765"
      city:
        type: "string"
        required: true
        example: "City"
      street:
        type: "string"
        example: "Street"
      streetNo:
        type: "string"
        example: "123"
      name:
        type: "string"
        example: "Company Building"
...
  User:
    type: "object"
    properties:
      id:
        type: "string"
        example: "12345"
      email:
        type: "string"
        example: "user@example.com"
      password:
        type: "string"
        example: "password"
      firstName:
        type: "string"
        example: "John"
      lastName:
        type: "string"
        example: "Doe"
      mn:
        type: "string"
        example: "987654321"
      birthplace:
        type: "string"
        example: "birthplace"
      companies:
        $ref: "#/definitions/UserCompany"
      phone:
        type: "string"
        example: "987654321"  
      addressId:
        type: "string"
        example: "54321"
      address:
        $ref: "#/definitions/Address"
      isDriver:
        type: "bool"
        example: "true"  
      categories:
        $ref: "#/definitions/DriverCategory"
      licenceNumber:
        type: "string"
        example: "12345"
      licenceSerialNumber:
        type: "string"
        example: "54321"
      licenceExpiry:
        type: "string"
        format: "date"
        example: "2023-12-31"
      licenceIssued:
        type: "string"
        format: "date"
        example: "2020-01-01"
      licenceAuthority:
        type: "string"
        example: "Authority"
      roles:
        type: "array"
        items:
          type: "string"
        example: [ "manager", "driver"]  
...
  DriverCategory:
    type: "object"
    properties:
      userId:
        type: "string"
        example: sdfsdfdsdfsdf-sdfsdfsdfdsf-sdfdfs-dfs
      category:
        type: string
        example: A
      issued:
        type: string
        format: date
        example: "2023-01-01"
      expired:
        type: string
        format: date
        example: "2024-01-01"
...
  Repair:
    type: "object"
    properties:
      warrantId:
        type: "warrantId"
        required: true
        example: "018cfd18-f674-7d5e-8051-a4dea1acbe79"
      warrant:
        $ref: "#/definitions/Warrant"  
      workshop:
        type: "string"
        required: true
        example: "workshop"
      repairType:
        type: "string"
        required: true
        example: "repairType"
      location:
        type: string
        example: 20,44
      addressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      address:
        $ref: "#/definitions/Address"
      start:
        type: "string"
        required: true
        format: "date-time"
        example: "2023-12-30 12:00:00"
      end:
        type: "string"
        format: "date-time"
        example: "2023-12-31 12:00:00"      
      fileName:
        type: string
        example: fiscal.bill.pdf
      url:
        type: string
        example: http://s3/temporartyUrl
  PatchRepairRequest:
    type: "object"
    properties:
      addressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      workshop:
        type: "string"
        example: "Updated workshop"
      repairType:
        type: "string"
        example: "Updated repairType"  
      start:
        type: "string"
        format: "date-time"
        example: "2023-12-30 12:00:00"
      end:
        type: "string"
        format: "date-time"
        example: "2023-12-31 12:00:00"
      location:
        type: "string"
        example: "20,44"  
      fileName:
        type: "string"
        example: "repair.revers.pdf"  
  CreateRepairRequest:
    type: "object"
    properties:
      addressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      warrantId:
        type: "string"
        example: "018cfd18-f674-7d5e-8051-a4dea1acbe79"
      workshop:
        type: "string"
        example: "workshop"
      repairType:
        type: "string"
        example: "repairType"
      start:
        type: "string"
        format: "date-time"
        example: "2023-12-30 12:00:00"
      end:
        type: "string"
        format: "date-time"
        example: "2023-12-31 12:00:00"
      fileName:
        type: string
        example: fiscal.bill.pdf
  TransportCosts:
    type: "object"
    properties:
      warrantId:
        type: "warrantId"
        example: "018cfd18-f674-7d5e-8051-a4dea1acbe79"
      type:
        type: "string" 
        example: "type"
      amount:
        type: number
        example: 23.2
      location:
        type: "string" 
        required: true
        example: 20,44
      code:
        type: "string" 
        example: "code"
      fileName:
        type: string
        example: fiscal.bill.pdf
      url: 
        type: string
        example: http://s3/temporartyUrl
  PatchTransportCostsRequest:
    type: "object"
    properties:
      type:
        type: "string" 
        example: "type"
      code:
        type: "string" 
        example: "Updated code"
      location:
        type: string
        example: 20,44
      fileName:
        type: string
        example: fiscal.bill.pdf
  CreateTransportCostsRequest:
    type: "object"
    properties:
      warrantId:
        type: "warrantId"
        example: "018cfd18-f674-7d5e-8051-a4dea1acbe79"
      type:
        type: "string" 
        example: "type"
      code:
        type: "string" 
        example: "code"
      location:
        type: "string" 
        example: 20,44
      amount:
        type: number 
        example: 232.2
      fileName:
        type: string"
        example: fiscal.bill.pdf
  UserCompany:
    type: object
    properties:
      isDeleted:
        type: bool
        example: false
      isActive:
        type: bool
      roles:
        type: "array"
        items:
          type: "string"
        example: [ "manager", "driver"] 
      companyId:
        type: string
        example:  safsdfsadfasd-fadsfsadffds-sdafasd
  PatchDriverCategoryRequest:
    type: "object"
    properties:
      category:
        type: "string"
        example: "Updated Category"
      issued:
        type: "string"
        format: "date"
        example: "2022-01-01"
      expired:
        type: "string"
        format: "date"
        example: "2023-01-01"
  CreateDriverCategoryRequest:
    type: "object"
    properties:
      userId:
        type: "string"
        example: "018cfd18-f5e6-7fe3-804e-2e9d8152de91"
      category:
        type: "string"
        example: "New Category"
      issued:
        type: "string"
        format: "date"
        example: "2022-01-01"
      expired:
        type: "string"
        format: "date"
        example: "2023-01-01"
...
  Warrant:
    type: object
    properties:
      id:
        type: string
        example: "1"
      issueDate:
        type: string
        format: date-time
        example: "2023-12-31 12:00:00"
      expectedStart:
        type: string
        format: date-time
        example: "2023-12-30 12:00:00"
      closingDate:
        type: string
        format: date-time
        example: "2023-12-31 12:00:00"
      driverId:
        type: string
        example: "018cfd18-f5e6-7fe3-804e-2e9d8152de91"
      driver:
        $ref: "#/definitions/User"
      passengers:
        type: array
        items: 
          type: string
        example: ["Vasily Zaitzev", "Luke Skywalker"]
      vehicleId:
        type: string
        example: "018cfd19-0da3-727a-806c-9293bb95c4cf"
      vehicle:
        $ref: "#/definitions/Vehicle"
      trailerId:
        type: string
        example: "018cfd19-0da3-727a-806c-9293bb95c4cf"
      trailer:
        $ref: "#/definitions/Vehicle"
      companyId:
        type: string
        example: "018cfd19-0d5b-7108-806f-78e1de9fbbcc"
      company:
        $ref: "#/definitions/Company"
      dispatcherId:
        type: string
        example: "018cfda7-310b-7f3a-800e-4d9369bd7818"
      dispatcher:
        $ref: "#/definitions/User"
      status:
        $ref: "#/definitions/WarrantStatus"
      name: 
        type: string
        example: redovna emezeta
      note:
        type: string
        example: vozac je imao zdravtvenih problema zadnje vreme
...
  Route:
    type: "object"
    properties:
      warrantId:
        type: "string"
        example: "018cfd18-f674-7d5e-8051-a4dea1acbe79"
      startMileage:
        type: number
        example: 3242342
      startAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"
      startAddress:
        $ref: "#/definitions/Address"
      startTime:
        type: "string" 
        example: "2022-01-01 15:04:05"
      endMileage:
        type: number
        example: 3242342
      endAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"  
      endAddress:
        $ref: "#/definitions/Address"
      endTime:
        type: "string"
        example: "2023-01-01 15:04:05"
      order: 
        type: number
        example: 1
      status: 
        type: string
        example: active
  PatchRouteRequest:
    type: "object"
    properties:
      startMileage:
        type: number
        example: 23323
      startAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"  
      startTime:
        type: "string" 
        example: "2024-01-01 15:04:05"
      endMileage:
        type: number
        example: 24422
      endAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"    
      endTime:
        type: "string"
        example: "2024-03-03 15:04:05"
      order:  
        type: number
        example: 3
      loadingInvoiceId:
        type: string
        example: assafdafs-sadfsdf-adsf-adfsfds
      unloadingInvoiceId:
        type: string
        example: assafdafs-sadfsdf-adsf-adfsfds
      status:
        type: string
        example: done
  CreateRouteRequest:
    type: "object"
    properties:
      warrantId:
        type: "string"
        example: "018cfd18-f674-7d5e-8051-a4dea1acbe79"
      startAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"  
      startTime:
        type: "string" 
        example: "2022-01-01 15:04:05"
      endAddressId:
        type: "string"
        example: "018cfd18-f492-7ade-803a-b414a0a8afac"    
      endTime:
        type: "string"
        example: "2023-01-01 15:04:05"
      order: 
        type: number
        example: 1
      status: 
        type: string
        example: ready
      loadingInvoiceId:
        type: string
        example: asdaddassd-as-dasdsadsadsad-dasads
      unloadingInvoiceId:
        type: string
        example: asdaddassd-as-dasdsadsadsad-dasads
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer